requirements.txt
beautifulsoup4==4.14.2
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
curl_cffi==0.13.0
dateparser==1.2.0
frozendict==2.4.6
html5lib==1.1
idna==3.11
joblib==1.4.2
llvmlite==0.44.0
markdown-it-py==4.0.0
matplotlib==3.9.0
mdurl==0.1.2
multitasking==0.0.12
newsapi-python==0.2.7
numba==0.61.2
numpy==1.26.4
packaging==24.2
pandas==2.2.2
peewee==3.18.2
pillow==10.4.0
platformdirs==4.5.0
protobuf==5.29.5
psycopg2-binary==2.9.9
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.5
rich==13.9.4
scikit-learn==1.5.2
six==1.17.0
soupsieve==2.8
sqlalchemy==2.0.34
streamlit==1.36.0
tenacity==8.5.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
vectorbt==0.26.2
watchdog==4.0.2
weasyprint==60.2
websockets==15.0.1
yfinance==0.2.66
PyYAML==6.0.1